Added psycopg2 library for integration with postgresql. Added a dump with an example of a catalog to be used as a test. DBBase class changes started. #1225
