+++ api/requirements.txt
@@ -31,3 +31,4 @@ celery==4.1.1
 coverage==4.5.2
 cx-Oracle==7.0.0
 defusedxml==0.5.0
+psycopg2-binary
